refactor(billing): neutralize usage and recharge service filenames
