+++ turbo/apps/api/src/signals/services/zero-credit-usage.service.ts
@@ -22,7 +22,7 @@ import {
   LOW_CREDIT_EMAIL_ALERT_THRESHOLD_CREDITS,
   type CreditLowBalanceAlertArgs,
 } from "./zero-credit-low-balance-alert.service";
-import { triggerAutoRecharge$ } from "./zero-credit-recharge.service";
+import { triggerAutoRecharge$ } from "./credit-recharge.service";
 import { applyUsageAllowanceToUsageEventsInLockedTransaction } from "./usage-allowance.service";
 import type { Tx } from "../../lib/db-types";
