+++ turbo/apps/api/src/signals/services/zero-usage.service.ts
@@ -9,8 +9,8 @@ import { userCache } from "@okouai/db/schema/user-cache";
 import { clerk$, type ClerkUser } from "../external/clerk";
 import { writeDb$ } from "../external/db";
 import { nowDate } from "../../lib/time";
-import { getOrgBillingPeriod$ } from "./zero-org-billing-period.service";
-import { getMemberUsageTotals } from "./zero-usage-reporting-ledger";
+import { getOrgBillingPeriod$ } from "./org-billing-period.service";
+import { getMemberUsageTotals } from "./usage-reporting-ledger";
 import { fixedRangeToPeriod } from "./usage-period";
 
 interface UsageMembersArgs {
